requirements.txt
# Pinned to versions used during development (Apr 2026).
# Tested on Python 3.11.

# --- Core data stack ---
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
pyarrow==16.1.0

# --- Modelling ---
scikit-learn==1.5.0
scikit-surprise==1.1.4

# --- Visualisation ---
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0

# --- Utilities ---
tqdm==4.66.4
PyYAML==6.0.1
joblib==1.4.2
boto3==1.35.53

# --- Pipeline / tests ---
pytest==8.3.3

# --- Notebook ---
jupyter==1.0.0
ipykernel==6.29.4
